Added RemappedName & ShellReplacementModder (Read description for info)
The combination of RemappedName and ShellReplacementModder allows for
the creation of "shell" classes. These can be given any name in their
source code, but with the RemappedName annotation can be given names
that will be used at runtime. A ShellReplacementModder must be used on
classes referencing these shells in order to do this name swapping.